Moved validator to own file. GetDebugScenario way more accurate now.
