+++ PowerShellTools.HostService.x86/PowerShellTools.HostService.x86.csproj
@@ -146,6 +146,9 @@
     <Compile Include="..\PowerShellTools.HostService\ServiceManagement\Debugging\PowerShellDebuggingService.cs">
       <Link>ServiceManagement\Debugging\PowerShellDebuggingService.cs</Link>
     </Compile>
+    <Compile Include="..\PowerShellTools.HostService\ServiceManagement\Debugging\PowerShellDebuggingServiceAttachValidator.cs">
+      <Link>ServiceManagement\Debugging\PowerShellDebuggingServiceAttachValidator.cs</Link>
+    </Compile>
     <Compile Include="..\PowerShellTools.HostService\ServiceManagement\Debugging\PowerShellDebuggingServiceEventHandlers.cs">
       <Link>ServiceManagement\Debugging\PowerShellDebuggingServiceEventHandlers.cs</Link>
     </Compile>
