+++ PowerShellTools.HostService/PowerShellTools.HostService.csproj
@@ -127,6 +127,7 @@
     <Compile Include="ServiceManagement\Debugging\DteManager.cs" />
     <Compile Include="ServiceManagement\Debugging\PowershellBreakpointRecord.cs" />
     <Compile Include="ServiceManagement\Debugging\PowerShellDebuggingService.cs" />
+    <Compile Include="ServiceManagement\Debugging\PowerShellDebuggingServiceAttachValidator.cs" />
     <Compile Include="ServiceManagement\Debugging\PowerShellDebuggingServiceEventHandlers.cs" />
     <Compile Include="ServiceManagement\Debugging\PowerShellDebuggingServiceUtilities.cs" />
     <Compile Include="ServiceManagement\Debugging\PowerShellDebuggingServiceUtilities40.cs" />
